Add "bxidle" option to parallel port's file output specification
This signifies that output should not be closed until bookexpert has
not output any new print data, and ALSO has been idle for at least
<timeout> milliseconds.  You must also specify the "timeout" option
whenever you use this "bxidle" option.

Typically three or five seconds would be enough to ensure that we're
not tricked into spooling the output prematurely.  This used to happen
because no print data would arrive and we'd assume the print job was
done; but Bookexpert was still working on producing it.